labs/lab02-intro-ml-networking/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
joblib>=1.3
requests>=2.31
netmiko>=4.3
pyyaml>=6.0
jupyter>=1.0
ipykernel>=6.0
